requirements.txt

google-generativeai
yfinance
fastapi
uvicorn
